requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-authtoken==2.1.0
numpy
opencv-python-headless
Pillow
tensorflow-cpu
torch
torchvision
gunicorn
whitenoise
requests
